+++ examples/svelte/src/routes/ui/components/authenticator/custom-slots/aws-exports.js
@@ -1,0 +1,2 @@
+import awsExports from '@environments/auth/auth-with-totp-mfa/src/aws-exports';
+export default awsExports;