+++ examples/svelte/src/routes/ui/components/authenticator/sign-up-with-email-lambda/aws-exports.js
@@ -1,0 +1,2 @@
+import awsExports from '@environments/auth/auth-with-email-lambda-signup-trigger/src/aws-exports';
+export default awsExports;