+++ packages/svelte/src/lib/components/Authenticator/index.ts
@@ -1,0 +1,29 @@
+import Authenticator from './Authenticator.svelte';
+import ConfirmResetPassword from './ConfirmResetPassword.svelte';
+import ConfirmSignIn from './ConfirmSignIn.svelte';
+import ConfirmSignUp from './ConfirmSignUp.svelte';
+import ConfirmVerifyUser from './ConfirmVerifyUser.svelte';
+import ForceNewPassword from './ForceNewPassword.svelte';
+import ForgotPassword from './ForgotPassword.svelte';
+import SelectMfaType from './SelectMfaType.svelte';
+import SetupEmail from './SetupEmail.svelte';
+import SetupTotp from './SetupTotp.svelte';
+import SignIn from './SignIn.svelte';
+import SignUp from './SignUp.svelte';
+import VerifyUser from './VerifyUser.svelte';
+
+export {
+  Authenticator,
+  ConfirmResetPassword,
+  ConfirmSignIn,
+  ConfirmSignUp,
+  ConfirmVerifyUser,
+  ForceNewPassword,
+  ForgotPassword,
+  SelectMfaType,
+  SetupEmail,
+  SetupTotp,
+  SignIn,
+  SignUp,
+  VerifyUser,
+};