+++ packages/svelte/src/lib/components/primitives/Label.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import { type HTMLLabelAttributes } from 'svelte/elements';
+
+  const { children, ...rest }: HTMLLabelAttributes = $props();
+</script>
+
+<label data-amplify-label {...rest}>
+  {@render children?.()}
+</label>