+++ packages/svelte/src/lib/components/primitives/Text.svelte
@@ -1,0 +1,9 @@
+<script lang="ts">
+  import { type HTMLAttributes } from 'svelte/elements';
+
+  const { children }: HTMLAttributes<HTMLSpanElement> = $props();
+</script>
+
+<span data-amplify-text="">
+  {@render children?.()}
+</span>