+++ packages/test-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @aws-amplify/ui-test-utils
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-storage@3.13.1
+
 ## 0.0.10
 
 ### Patch Changes